# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Referral Robin API — fair-rotation referral code marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Brand {
  id: number;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  currentOffer?: string | null;
  /** @nullable */
  offerUpdatedAt?: string | null;
  category: string;
  active: boolean;
  /** Number of active codes in the queue */
  codeCount?: number;
}

export interface CodeReveal {
  codeId: number;
  code: string;
  brandId: number;
  brandName: string;
  /** @nullable */
  currentOffer?: string | null;
  cooldownUntil: string;
}

export interface NextCodeInput {
  brandId: number;
  deviceId: string;
}

export interface ConfirmCopyInput {
  codeId: number;
  deviceId: string;
}

export interface CodeInput {
  brandId: number;
  /** @minLength 1 */
  code: string;
}

export interface ReportInput {
  deviceId: string;
}

export type UserCodeStatus = typeof UserCodeStatus[keyof typeof UserCodeStatus];


export const UserCodeStatus = {
  active: 'active',
  paused: 'paused',
  removed: 'removed',
} as const;

export interface UserCode {
  id: number;
  brandId: number;
  brandName: string;
  code: string;
  status: UserCodeStatus;
  timesServed: number;
  timesCopied: number;
  createdAt: string;
}

export interface UserStats {
  totalCodesSubmitted: number;
  totalTimesServed: number;
  totalTimesCopied: number;
}

export interface CooldownStatus {
  isOnCooldown: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  remainingSeconds?: number | null;
}

export interface CooldownError {
  error: string;
  expiresAt: string;
  remainingSeconds: number;
}

export interface SuccessResult {
  success: boolean;
}

export type ListBrandsParams = {
category?: string;
search?: string;
};

export type GetCooldownParams = {
brandId: number;
deviceId: string;
};