# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Referral Robin API — fair-rotation referral code marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface Brand {
  id: number;
  name: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  currentOffer?: string | null;
  /** @nullable */
  offerUpdatedAt?: string | null;
  category: string;
  active: boolean;
  /** Number of active codes in the queue */
  codeCount?: number;
}