# lib/api-zod/src/generated/types/cooldownStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Referral Robin API — fair-rotation referral code marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface CooldownStatus {
  isOnCooldown: boolean;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  remainingSeconds?: number | null;
}