# lib/api-zod/src/generated/types/userCode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Referral Robin API — fair-rotation referral code marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { UserCodeStatus } from './userCodeStatus';

export interface UserCode {
  id: number;
  brandId: number;
  brandName: string;
  code: string;
  status: UserCodeStatus;
  timesServed: number;
  timesCopied: number;
  createdAt: Date;
}